+++ index.js
@@ -1,3 +1,5 @@
+import Money from './src/money.mjs'
+
 window.addEventListener('load', () => {
   // const params = new Proxy(new URLSearchParams(window.location.search), {
   //   get: (searchParams, prop) => searchParams.get(prop),
@@ -14,68 +16,22 @@ window.addEventListener('load', () => {
 
   let mOH = new Money(localStorage.getItem('moneyOnHand') || 0, true)
   document.getElementById('moneyOnHand').value = mOH.toString()
-  document.getElementById('adjustMoney').onkeyup = handleKeyup
-  document.getElementById('category').onkeyup = handleKeyup
   initHistoryList()
   populateCategories(localStorage.getItem('categories'))
 
   if ('serviceWorker' in navigator) {
     navigator.serviceWorker.register('./src/sw.js')
   }
+
+  document.getElementById('adjustMoney').addEventListener('input', checkAmount)
+  document.getElementById('moneyOnHand').addEventListener('blur', updateMoneyOnHand)
+
+  document.getElementById('moneyForm').addEventListener('submit', e => {
+    e.preventDefault()
+
+    adjustMoney()
+  }, false)
 })
-
-class Money {
-  #amount
-  #formatter
-  constructor(amount, asIs = false, locale = 'en-US', currency = 'USD') {
-    this.#amount = asIs ? amount : amount * 100
-    this.#formatter = new Intl.NumberFormat(locale, { style: 'currency', currency })
-  }
-
-  static fromFormatted(amount) {
-    return new Money(
-      amount.padEnd(
-        amount.lastIndexOf('.') + 3,
-        '0')
-        .replace(/[,\._]/g, ''),
-      [...amount.matchAll(/\./g)].length !== 0)
-  }
-
-  static Zero = new Money(0)
-
-  #isEqual(that) {
-    return Math.abs(this.#amount - that.#amount) < Number.EPSILON
-  }
-
-  toString(showCurrency = false, numOnly = false, dec = true) {
-    // force showCurrency false if numOnly true to make switch a little easier
-    if (numOnly) {
-      showCurrency = false
-    }
-    return this.#formatter.formatToParts(Math.round(this.#amount) / 100).map(({ type, value }) => {
-      switch (type) {
-        case 'currency': return showCurrency ? value : ''
-        case 'decimal': return dec ? value : ''
-        case 'group':
-        case 'infinity':
-        case 'literal':
-        case 'nan':
-          return numOnly ? '' : value
-        default: return value
-      }
-    }).reduce((string, part) => string + part)
-  }
-
-  add(that) {
-    this.#amount += that.#amount
-    return this
-  }
-
-  sub(that) {
-    this.#amount -= that.#amount
-    return this
-  }
-}
 
 function initHistoryList() {
   let histList = document.getElementById('historyList')
@@ -182,11 +138,5 @@ function checkAmount() {
     }
   } else {
     amount.value = lastValidAmount
-  }
-}
-
-function handleKeyup(e) {
-  if (e.key === "Enter") {
-    adjustMoney()
   }
 }
