+++ style.css
@@ -26,7 +26,7 @@ input:invalid {
   align-items: center;
 }
 
-.content {
+.content > * {
   color: #444;
   display: flex;
   flex-direction: column;
